+++ babel-presets/react-prod.js
@@ -1,6 +1,0 @@
-module.exports = {
-  plugins: [
-    require('babel-plugin-transform-react-constant-elements'),
-    require('babel-plugin-transform-react-remove-prop-types').default
-  ]
-}