+++ src/moduleBuild.js
@@ -138,9 +138,9 @@ export default function moduleBuild(args, buildConfig = {}, cb) {
       'ES5',
       {copyFiles, outDir: path.resolve('lib'), src},
       merge(buildConfig.babel, buildConfig.babelDev || {}, {
-        // Don't force ES5 users of the ES5 build to eat a .require
+        // Don't force CommonJS users of the CommonJS build to eat a .require
         commonJSInterop: true,
-        // Transpile modules to CommonJS for ES5 users
+        // Transpile modules to CommonJS
         modules: 'commonjs',
         // Don't set the path to nwb's babel-runtime, as it will need to be a
         // peerDependency of your module if you use transform-runtime's helpers